requirements.txt
aiofiles==0.5.0
aiosqlite==0.15.0
aniso8601==8.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.21.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.0
graphene==2.1.8
graphql-core==3.1.2
graphql-relay==3.0.0
h11==0.9.0
httptools==0.1.1
idna==2.10
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==3.3.0
promise==2.3
pydantic==1.6.1
PyPika==0.38.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
Rx==3.1.1
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.14
typing-extensions==3.7.4.2
ujson==3.1.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
